refactor: add custom hook for likes
