+++ src/screens/Store/StoreDetailScreen/StoreDetailScreen.tsx
@@ -10,12 +10,12 @@ import {
 } from 'react-native';
 import { useLocalSearchParams, useRouter } from 'expo-router';
 import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { LinearGradient } from 'expo-linear-gradient';
 import { getStoreDetail } from '../../../api/store';
 import { getReviewCount, getReviewPreviews } from '../../../api/review';
-import { getIsLiked, likeStore, unlikeStore } from '../../../api/shop';
-import useShopStore from '../../../store/useShopStore';
+import { getIsLiked } from '../../../api/shop';
+import useLikeStore from '../../../hooks/useLikeStore';
 import type { StoreDetail } from '../../../types/store';
 import styles from './StoreDetailScreen.style';
 import ReviewCard from '../ReviewCard/ReviewCard';
@@ -40,36 +40,13 @@ function StoreDetailScreen() {
   const insets = useSafeAreaInsets();
   const [headerH, setHeaderH] = React.useState(0);
   const [HeaderS, setHeaderS] = React.useState(0);
-  const queryClient = useQueryClient();
 
   const { id, from: rawFrom } = useLocalSearchParams<{ id?: string; from?: string | string[] }>();
   const from = Array.isArray(rawFrom) ? rawFrom[0] : rawFrom;
   const storeId = id;
-  const { addFavoriteShop, removeFavoriteShop, isFavoriteShop } = useShopStore();
+  const { toggleLike, isFavoriteShop, addFavoriteShop, removeFavoriteShop } = useLikeStore();
 
   const isLiked = storeId ? isFavoriteShop(storeId) : false;
-
-  const { mutate: like } = useMutation({
-    mutationFn: () => likeStore(storeId!),
-    onMutate: async () => {
-      await queryClient.cancelQueries({ queryKey: ['isLiked', storeId] });
-      addFavoriteShop(storeId!);
-    },
-    onError: () => {
-      removeFavoriteShop(storeId!);
-    },
-  });
-
-  const { mutate: unlike } = useMutation({
-    mutationFn: () => unlikeStore(storeId!),
-    onMutate: async () => {
-      await queryClient.cancelQueries({ queryKey: ['isLiked', storeId] });
-      removeFavoriteShop(storeId!);
-    },
-    onError: () => {
-      addFavoriteShop(storeId!);
-    },
-  });
 
   // 상세 데이터 패칭
   const {
@@ -151,15 +128,6 @@ function StoreDetailScreen() {
     else router.back();
   };
 
-  const handleLike = () => {
-    if (!storeId) return;
-    if (isLiked) {
-      unlike();
-    } else {
-      like();
-    }
-  };
-
   const details = [
     { label: '카테고리', value: store.categoryName },
     { label: '매장번호', value: store.merchantNumber || '-' },
@@ -203,7 +171,7 @@ function StoreDetailScreen() {
             <TouchableOpacity
               onPress={(e) => {
                 e.stopPropagation?.();
-                handleLike();
+                if (storeId) toggleLike(storeId);
               }}
               hitSlop={{ top: 10, bottom: 10, left: 10, right: 10 }}
             >
